Fix confirmation of emails when the submit button is image
Problem
-------

Confirmation of email does not work when the submit button is image.

Suppose that we have a form having the following fields:

* Email
    * type: E-Mail
    * required: true
    * confirm title: "Confirm Email"
* Confirm Email
    * type: E-Mail
    * required: true
* Submit Button
    * Use button image

When a user enters two different emails and then click the submit button,
they see no error and proceed to the confirm screen.

Expected Behavior
-----------------

When the user click the submit button, they see an error message
"The e-mail mismatch".

Solution
--------

Both send-to-confirm and send-to-confirm_x should be checked.